almost refactored testAddingData.py
